Add Video Thumbnail With Black Bars Padding
Uses FFmpeg to get one frame to produce the thumbnail. This
thumbnail uses the dynamically passed width + height to produce
the required different thumbnail sizes and uses black to fill the
space.

This is used in combination with no longer making all videos
square.

This commit will allow the gallery to still have regimented thumbnails
for preview, but then the videos original aspect ratio for playback.